move the info-button icon into an actual button
By wrapping the svg into a button, a keyboard user can access it, but it
is more explicit in what it means. this thing functions as a button, so
make it one. We could (should) also add text in the button 'more info'
or something and hide that for visual users