set all children inside info-button to not receive pointer events
We set the click handler on the button, but the ACTUAL clicked element
will be the path IN the svg IN the button. By using pointer-events: none
on all the children of the button, they will not respond to mouse
events. Hence the button element will show up as event.target