Move styles.css to resources/styles/main.css
We need a single entry point for our css. This makes it (later on)
easier to follow which css is included in what html file. And even
further down the line, we are going to minify the css. To prepare for
that we need a single css file to start from