Move each selector to its own line
Although these selectors are quite small and the do fit on a line, it is
usually best to place them on seperate lines. When selectors become
larger / more complex, having them on its own line makes it more
readable. Since that will be the case later on, consistency is king and
we should start doing that right now