Move tip-box styles into own component
Although the tip-box is only used in the ControlPanel, it is a self
contained piece of functionality that could have been used elseware. As
such, it makes sense to make it into it's own component. It is
debateable if the trigger (the icon in this case) should be part of it
as well, but I figured that the trigger is an independent action. And
thus it is NOT included in the component