Remove all example text from html
In the future we want to make it possible to add/remove classes. That
means that we also need to add/remove example texts. Thus, we are going
to create and add the elements via JavaScript. We add the data-content
attribute so we can targer the content element more easy