Toggle the control panel with data attributes
We set the state of the control panel via javascript. In this case via
the data-control-panel="open" or "close" attribute. It is convention to
use data attributes over classes because now it it clear from the html
that these values are modified by JavaScript. Also if you look at the
css, you see which styles are controlled by javascript because of the
[data-*=""] syntax

Further more, we use an actual button instead of a div. This way we get
keyboard interactivity for free! (once we apply focus styling...)

note that I did include the javascript file, in this commit, but it is
not used yet. In a future commit this will be include.