We create a control panel script wich handles controlpanel interactions
For now, only the Type Scale section is targeted, but the code is made
so generic that the Base Type Size should be really easy to implement.
See next commit.
The code is heavily commented so you can see what is going on.

New CSS is added for the new structure of the input elements. That makes
it easier to understand rather than modifying all existing css.