Generate the Base Type controls from JavaScript
First we remove the hardcoded controls from the HTML. These are going to
be generated via JavaScript.

Then we generate them in control-panel.js. All the confusing generic
functions in control-panel.js start to make a little sense. Because we
made the functions generic, we can reuse them for a completely new form
block. Notice the amount of JavaScript changes in control-panel.js in
this commit. Now imagine that, in the future, we need to add yet another
form block which is dependend on the active breakpoint. That would be
really easy.