Generate the breakpoint icons and inputs via JavaScript
For the same reason we generate all breakpoint-dependent controls via
JavaScript, if the user adds a breakpoint we need to manage the DOM, we do the same
for these. The added function takes two wrappers, each get filled with the icons and inputs.

Notice that we can reuse the function to highlight the active
breakpoint!

In the CSS we show/hide the images depending on the data attributes
which are set via JavaScript.