Generate limit controls via JavaScript
It is likely that the user wants to set their own limits for specific
classes in the future. Therefor, we should generate them via JavaScript.
We can reuse the same code that we have written for the basetype and
typescale controls, except that the limits are not breakpoint dependend.

This makes the naming of the functions which create the controls to
specific. It includes `sizes`. But now we also want them to function
with a different data set. So we rename the functions to be more
generic, as the classnames of the elements created.

Lastly we need to modify the css to set the opacity of input elements
only if it is breakpoint dependend (if it has a data-breakpoint
attribute)