Display output list via JavaScript
Again, a lot of refactoring here. The output list has quite a lot of
functionality in common with the input elements:
* It is based on a list in `script-new.js`
* It needs to render based on the classes the user is going to provide
* It is a list of labels and values

Actually the only thing really different is the kind of element used to
display the data, `input` for almost everything, `output` for this list.

We already have a specific function which generates the actual elements.
That is the function in lib/create-control-elements.js. The function
displayControlElements in control-panel.js only loops over an array and
invokes that function... In JavaScript we can pass around functions as
if it were values. So if displayControlElements does not create the
actual element itself, we can pass in the function that does as an
argument. Well, now the name of displayControlElements is no longer
correct. It can do much more. So we rename it to be a bit more generic,
as well as some variables inside of it

Then we create a new function in lib/create-control-element.js to
generate an output element. We rename the file because it does not deal
only with control elements anymore.

Because we refactored the displayFormElements function to accept a
function which does the actual creation, we need to pass that in within
setup.

Lastly we modify the HTML and pass the classes array into control-panel.
One thing to note in script-new.js is that there is now a dummy function
which binds the classname to a (for now) static value. That is something
we will come back to in the future