Generate the breakpoint indicators in the background with JavaScript
The user can, in the future, add breakpoints. So they need to be
generated via JavaScript. The code is relatively the same as content.js
and control-panel.js.

We export a function which returns an object. That object has a function
updateBreakpoints to be called by script-new.js. When the new sizes
array comes in, it generates breakpoint elements for each size except
the Infinity one.

The css for it is modified, because we don't know how much elements
there will be, we can't hardcode them all. So a class is used