Create update functions for each variable in script-new.js
Various components want to be notified of the variables in
script-new.js. If we want to change one of the variables, we need to
remember to call all components which depend on that variable, as the
application scales, this becomes harder to remember. Now if we update a
variable, we only need to call a single function.