Generate the based on sizes object
The css is generated in generate-css.js. This function results in a
string which is nicely formatted, ready to be pasted css. It mostly
loops over arrays and constructs the text, adds some indentation and
places comments with the px value. All values are based on rems, instead
of hard px values, to allow for zooming by the user. Because this
function outputs a string of css, we can use this exact function to
create a download link for the user later on.

The head-css.js file merely puts the css in to a style tag. Because we
no longer put the values inline on the example text elements, we now use
the exact same css on the site as the user would download it from us!

We wire up everything in script-new.js like we did with other variables
as well