Respond to user by recalculating sizes and update values
We create event handlers for each form group and attach them when the
elements are created. We need to preform some complex things because we
need to keep a reference to the actual event handlers in order to remove
them once the input elements are destroyed.

The script-new.js file becomes a bit complex. Like before it has
functions for setting variables (setSizes, setLimits), but if either of
these are executed, almost all form blocks need to be updated in the
right order. This was exactly what I wanted to avoid, but here we are...

In the future, we need to revisit this architecture and try to keep the
input elements instead of destroy/readd them on every change.