Keep existing input elements when updating
I found that when the user modifies a value, it impacts nearly all parts
of the application. That meant that every input element was recreated
everytime the user changes something. However, the script-new.js file
was setup to only update the parts that changed. However, not only did
all parts update, there was a specific order to update. That made it all
quite complicated.

In script-new.js:
Now, I made a seperation between code sets variables and code that
updates. I got rid of all update*() functions, everything needs to
update anyway. We now have calculate functions (calculating the
dependencies) and set functions setting variables. when those run in
order, then we are ready for the update call.

In control-panel.js:
The createElementFunctions now accept an existing element, when found
these are reused.
I also moved the creation of the breakpoint elements into the
create-element.js file, this makes control-panel.js more readable